backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.4
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
openai==1.3.7
httpx==0.25.2
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 